requirements.txt
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asyncpg==0.30.0
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
black==24.10.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
fake-useragent==2.0.3
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
lxml==5.3.0
lxml_html_clean==0.4.1
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
pathspec==0.12.1
platformdirs==4.3.6
playwright==1.49.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
requests-html==0.10.0
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.1
w3lib==2.2.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
